+++ src/ReduxToastr.js
@@ -1,40 +1,51 @@
-import React,
-  {Component, PropTypes}    from 'react';
+import React, {Component, PropTypes} from 'react';
 import {connect}            from 'react-redux';
 import {bindActionCreators} from 'redux';
 import classnames           from 'classnames';
 
 import ToastrBox            from './ToastrBox';
+import ToastrConfirm        from './ToastrConfirm';
 import * as tActions        from './actions';
 import {EE}                 from './toastrEmitter';
 import config               from './config';
 
-import {checkPositionName, isMobile}  from './utils.js';
+import {checkPositionName, isMobile, _bind, hasProperty} from './utils.js';
 
 const mapStateToProps = (state) => ({
-  toastrs: state.toastr
+  toastr: state.toastr
 });
 
 export class ReduxToastr extends Component {
   static displayName = 'ReduxToastr'
 
   static propTypes = {
-    toastrs: PropTypes.array,
+    toastr: PropTypes.object,
+    options: PropTypes.object,
     position: PropTypes.string,
     newestOnTop: PropTypes.bool,
-    timeOut: PropTypes.number
+    timeOut: PropTypes.number,
+    confirm: PropTypes.object
   }
 
   static defaultProps = {
     position: 'top-right',
-    newestOnTop: true
+    newestOnTop: true,
+    confirm: {
+      okText: 'ok',
+      onCancelText: 'cancel'
+    }
   }
 
   constructor(props) {
     super(props);
     this.actions = bindActionCreators(tActions, this.props.dispatch);
     config.set('newestOnTop', this.props.newestOnTop);
-    this.handleRemoveToastr = this.handleRemoveToastr.bind(this);
+
+    _bind(
+      this,
+      'handleRemoveToastr',
+      'handleHideConfirm'
+    );
   }
 
   componentDidMount() {
@@ -43,16 +54,24 @@ export class ReduxToastr extends Component {
     };
 
     const onCleanToastr = () => {
-      if (this.props.toastrs.length) {
+      if (this.props.toastr.toastrs.length) {
         this.actions.clean();
       }
     };
+    const confirm = (obj) => {
+      // Fire if we don't have any active confirm
+      if (!this.props.toastr.confirm.show) {
+        this.actions.confirm(obj.message, obj.options);
+      }
+    };
 
+    EE.on('toastr/confirm', confirm);
     EE.on('add/toastr', onAddToastr);
     EE.on('clean/toastr', onCleanToastr);
   }
 
   componentWillUnmount() {
+    EE.removeListener('toastr/confirm');
     EE.removeListener('add/toastr');
     EE.removeListener('clean/toastr');
   }
@@ -61,17 +80,31 @@ export class ReduxToastr extends Component {
     this.actions.remove(id);
   }
 
+  handleHideConfirm() {
+    this.actions.hideConfirm();
+  }
+
   render() {
     const toastrPosition = checkPositionName(this.props.position);
     const classes = classnames('redux-toastr', toastrPosition, {mobile: isMobile});
+    const {toastr, confirm} = this.props;
+
+    const confirmOkText = hasProperty(confirm, 'okText') ? confirm.okText : 'ok';
+    const confirmCancelText = hasProperty(confirm, 'cancelText') ? confirm.cancelText : 'cancel';
 
     return (
       <div className={classes}>
-        {this.props.toastrs.map((toastr) => {
+        <ToastrConfirm
+          hideConfirm={this.handleHideConfirm}
+          confirm={toastr.confirm}
+          okText={confirmOkText}
+          cancelText={confirmCancelText}/>
+
+        {toastr.toastrs.map((item) => {
           return (
             <ToastrBox
-              key={toastr.id}
-              toastr={toastr}
+              key={item.id}
+              toastr={item}
               timeOut={this.props.timeOut}
               remove={this.handleRemoveToastr}/>
           );