+++ src/constants.js
@@ -1,3 +1,5 @@
-export const ADD_TOASTR     = 'reduxToastr/ADD_TOASTR';
-export const REMOVE_TOASTR  = 'reduxToastr/REMOVE_TOASTR';
-export const CLEAN_TOASTR   = 'reduxToastr/CLEAN_TOASTR';
+export const ADD_TOASTR     = '@ReduxToastr/ADD_TOASTR';
+export const REMOVE_TOASTR  = '@ReduxToastr/REMOVE_TOASTR';
+export const CLEAN_TOASTR   = '@ReduxToastr/CLEAN_TOASTR';
+export const SHOW_CONFIRM   = '@ReduxToastr/SHOW_CONFIRM';
+export const HIDE_CONFIRM   = '@ReduxToastr/HIDE_CONFIRM';