+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+#1.0.0
+##### add confirm method
+
 #0.4.0
-####add method message
-####add option newestOnTop
-####add mobile style
+#### add method message
+#### add option newestOnTop
+#### add mobile style
 
 #0.3.1
 Nothing new just some problems with npm readme.md file